+++ E_paper.h
@@ -7,9 +7,9 @@
 //  if you are not sure which version, please test each one,
 //  if it is successful then it belongs to the model of the file name
 // The 2.66 screen has other SPI pins, and more pixels. Partof the pixels are now unused.
-#include <GxDEPG0213BN/GxDEPG0213BN.h>  //Huidige schermen
+//#include <GxDEPG0213BN/GxDEPG0213BN.h>  //Huidige schermen
 //#include <GxGDEH0213B73/GxGDEH0213B73.h>  //Eerste schermen
-//#include <GxGDEM0213B74/GxGDEM0213B74.h>  // 2.13", opgepast hier rst en busy aanpassen 
+#include <GxGDEM0213B74/GxGDEM0213B74.h>  // 2.13", opgepast hier rst en busy aanpassen 
 //#include <GxDEPG0266BN/GxDEPG0266BN.h>  //2.66 screen for rollz...
 //#include <GxGDEW0213M21/GxGDEW0213M21.h>  //nog een andere versie, resolutie NIET OK vanwege resolutie !!!
 //#include <GxGDEP015OC1/GxGDEP015OC1.h>    //  GDEH015OC1 1.54" screen for Jeff Turner !!!
@@ -68,6 +68,7 @@ Flip: horizontally
 #define WIFI_STATION 12
 #define WIFI_SOFT_AP 13
 #define SPEED2 14
+#define TROUBLE 15
 #define MINIMUM_VOLTAGE 3.1       // if lower then minimum_voltage, back to sleep.....
 
 extern GxEPD_Class display;
@@ -83,6 +84,9 @@ extern int bootCount,run_count,stat_count,GPIO12_screen;
 extern int ftpStatus;
 extern String IP_adress;
 extern const char SW_version[16];
+extern UBXMessage ubxMessage;
+static int update_epaper=2;
+extern int freeSpace;
 extern RTC_DATA_ATTR int offset;
 extern RTC_DATA_ATTR float RTC_distance;
 extern RTC_DATA_ATTR float RTC_avg_10s;
@@ -119,6 +123,7 @@ void Update_screen(int screen);
 void Bat_level(int X_offset,int Y_offset);
 void Bat_level_Simon(int offset);
 void Sat_level(int offset);
+void time_print(int time);
 class Button_push{
             public:
             Button_push(int GPIO_pin,int push_time,int long_pulse_time,int max_count);//constructor